Implements CRUD for common Create/Read/Update/delete actions
Implements a default CRUD controller to handle all create, read, update and delete actions.
Client applications, properties and videos are now entities implementing CRUD operations. CRUD operations are handled by the openveo-core for all plugins. An entity is associated to a model which is associated to a provider. Distinction is now made between controllers, models and providers. This leads to externalize UserProvider, PropertyProvider and VideoProvider from openveo-api.