Roles / users / permissions implementation
Add two entities : "users" and "roles".
A user is defined by a name, an email, a password and one or several roles. User of id 0 is considered as the main administrator of the back end and can't be removed or modified using the interface.
A role is defined by a name and a list of permissions.
Permissions are defined in configuration files and represent access to paths.
It's not possible, anymore, to log to the back end with user name, email is now used.