Reorganize routes and public resources
All public routes corresponding to the back end (/authenticate, /logout, /getMenu etc.) are now mounted under /be and the root of the AngularJS single page application is /be instead of /.
Routes are now loaded as followed :

/ (assets)
/
/be

/plugin1 (assets)
/plugin1
/be/plugin1

/plugin2 (assets)
/plugin2
/be/plugin2

etc.

"public" directory has been renamed to "assets" and contains resources organized by functionalities.