Correct server error on errorController.errorAction and offer the possibilty to add an error message
When a JavaScript error happened through an express route, the error was forwarded to the errorController but the error wasn't handled properly, generating a server error. It now treats the error as a general server error.
It is now possible to add a message to errors defined in httpErrors using "message" property.