Add script to create a super admin
Super admin used to be created using the installation script at
project's installation. It was prompting the user for email, password
and super admin name.
Now, the installation script works exactly the same but super admin
creation has been moved into a separate script which makes it possible
to call it directly with name, email and password as options. It is a
non blocking script.