Add a directive and a formly field to match values
The directive ov-match and its corresponding formly type "match" is able to associate a value with a list of values. It is made of two fields: an input field holding a simple value and a tags field (using directive ov-tags) holding a list of values. Depending on its configuration several associations can be added using one directive.