Add support for CAS and LDAP
OpenVeo can now authenticates users using external services (CAS and LDAP).
A settings page on the back end has been added to configure CAS and LDAP users. Through this page it is possible to associate CAS / LDAP groups to OpenVeo roles.
Each user authenticated through CAS and / or LDAP is created on OpenVeo.
A CAS / LDAP user can't modify its profile but profile page is still available to display some basic information. Also the origin of the user is displayed in the list of users with the possibility to search by origin.