Create a user for Web Service clients
Requests authenticated through the Web Service
have an associated client with an id and a list of
scopes but didn't have a user as created by
passport for back office authentication. As Web
Service scopes only controls URLs, a request user
has to be set to perform access control on other
permissions. For example, in the new version of
@openveo/api, the
ContentController.isUserAuthorized() method
returns false if user is not specified, which was
the case for requests authenticated through the
Web Service. A user is now associated to requests
authenticated through the Web Service will full
privileges.