Remove formly label / field association
Formly fields were linked to their label by the
"for" attribute. This was causing unexpected
behaviour when switching between the editable
version of the field and the readable version of
the field. For example it was possible to change
the value of a checkbox for a field of type
editableCheckbox. Label / field association has
been removed as formly contains bugs which prevent
from disabling / enabling a field when the form
switches between readble mode and editable mode.