Add dateStart and dateEnd to search date filter
"dateStart" and "dateEnd" can be used as values
for a tableContainer filter of type "date".
"dateStart" will search for records starting at
the given date while "dateEnd" will search for
records ending at the given date. "date" value
still searches on the given date. With
"dateStart" and "dateEnd" it is possible to filter
records by a dates range.