Add parameters support in permissions rules
Some paths match ExpressJS routes containing
parameters, but it wasn't possible to authorize
these paths without making use of the wildcard
character in permissions / scope rules.
For example, to authorize the path
"/path/with/param" (with "param" a dynamic value)
the only way was to build a rule like
"/path/with/*". It worked but had limitations. For
example to authorize the path
"/path/with/param/and/action" (with "param" a
dynamic value) we had to add the rule
"/path/with/*" but what if we don't want to
authorize paths like
"/path/with/param/and/another/action"?

It is now possible to authorize paths with
parameters by declaring the parameter in the rule
like ExpressJS does. For example to authorize the
path "/path/with/param/and/action" and not the
path "/path/with/param/and/another/action" it is
possible to add the rule
"/path/with/:param-placeholder/and/action".