Make tableForm filterBy property more consistent
The tableForm filterBy property was sometimes
using the "key" property as the web service
parameter but sometimes it was the "param"
property. Filters of type "text" were merged to
build the "query" parameter ignoring the "key"
property. Finally the "filterWithChildren" was a
mess.
Everything has been simplified, the "type"
property doesn't have any impact on the value, the
"param" and the "filterWithChildren" property have
been removed, the "key" property now always refers
to the web service parameter and a new property
"getValue" can hold a function to act on the value
before sending it to the web service.
Plugins will have to remove their "param"
property, modify their "key" property and add a
"getValue" property if they were using "param"
property with values "date", "dateStart" or
"startEnd", or if they were using the property
"fillWithChildren".