Add CSS classes to table filters
Filters are displayed on 4 columns (on medium
screens and above) making the fifth filter and
the next ones to be displayed on the next lines
and it wasn't possible to select another line
of filters than the first one.
New CSS classes have been added to be able to
distinguish first line filters and next filters.