doc: add command to get specific chrome driver
The latest version of the Chrome Driver that the
webdriver manager can get is not always the good
one for the installed version of Google Chrome.
The precise version of Google Chrome can be
precised to the webdriver manager to get the
corresponding version of the Chrome Driver. This
command has been added to the documentation.