client.admin: fix authentication with replicasets
Login to the administration interface was working
but user wasn't redirected to the administration
interface after being authenticated when OpenVeo
is installed with MongoDB replicasets.
MongoDB does not have the time to replicate the
write operation on the core_sessions collection
holding the user session and the MongoStore which
manages the core_sessions does not set read
preferences on the collection to avoid reading on
a secondary.
The best solution will be to notify the maintainer
of the connect-mongo package to add the
possibility to set the read preference on the
collection.
Another option will be to create our own session
store.
Until then a delay has been added to the client to
wait for 100ms before trying to redirect to the
administration page.