tests: fix test on login page fields
Test "Login page should no consider fields on
error when modified" was failing because the
LoginPage.unfocus() action was unfocusing fields
by clicking on the center of the body element
which appears to be the submit button of the
formular.
No unfocus action is needed anymore because the
unfocus action is performed automatically by
e2e.fields.TextField.